fix: generate tunnel id only if starting sauce connect
When tunnelIdentifier is not explicitly configured, generate a random id only if starting sauce connect.

When the developer does not want to start sauce connect, it's more likely he is reusing some already running instance of sauce connect and it's more likely this instance is not using any tunnel id.