Allow connections through Sauce Connect's Selenium Relay.
Sauce Connect can function as a Selenium command relay, helping avoid proxy issues.
By default, the behaviour is not changed.  However, users can set:

`connectLocationForSERelay` to the host location of Sauce Connect

`connectPortForSERelay` to the port used to run Sauce Connect's Selenium relay.
This will default to the value set for `port` in `connectOptions`, then to `80`.