fix: do not convert IE9 from JWP to W3C
Update the automatic conversion of capabilities from JWP to W3C format
to exclude Internet Explorer 9. Although according to the
[SauceLabs docs][1] all versions of Internet Explorer support the new
W3C capabilities format, in reality only versions >=10 do.

SauceLabs support has confirmed this and said that until the docs are
fixed, _"the [Platform Configurator][2] would be the most reliable
reference for compliant capabilities"_. Indeed, in the Platform
Configurator, IE 9 only offers the option to select `JWP (legacy)` as
the capabilities format.

[1]: https://docs.saucelabs.com/dev/w3c-webdriver-capabilities/index.html#browser-compatibility
[2]: https://saucelabs.com/platform/platform-configurator