fix: use `appium:platformVersion` in `browserName` (if present)
Use the value of `appium:platformVersion` (if present) in `browserName`.

Since the [capabilities config when using Appium][1] does not include a
browser version, but only the platform version, including the value of
`appium:platformVersion` in `browserName` is essential in order to be
able to identify which platform the logs refer to.

[1]: https://docs.saucelabs.com/mobile-apps/automated-testing/appium/virtual-devices/#ios-code-examples